# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bluntly.ph API",
    "description": "Verified product & seller review platform. All error responses use the RFC 9457 problem+json schema. Identity is a FastAPI-issued JWT (ADR-010).",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/users/{user_id}/trust": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Public trust profile (stage, reputation, badges)",
        "operationId": "get_trust_api_v1_users__user_id__trust_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserTrustOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{user_id}/role": {
      "patch": {
        "tags": [
          "users"
        ],
        "summary": "Promote/demote seller (moderator only)",
        "operationId": "set_role_api_v1_users__user_id__role_patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sellers/{seller_id}/reviews": {
      "post": {
        "tags": [
          "sellers"
        ],
        "summary": "Review a seller (one per reviewer; publishes immediately)",
        "operationId": "create_seller_review_api_v1_sellers__seller_id__reviews_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "seller_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Seller Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SellerReviewCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SellerReviewOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "sellers"
        ],
        "summary": "List a seller's reviews (newest first)",
        "operationId": "list_seller_reviews_api_v1_sellers__seller_id__reviews_get",
        "parameters": [
          {
            "name": "seller_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Seller Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SellerReviewOut"
                  },
                  "title": "Response List Seller Reviews Api V1 Sellers  Seller Id  Reviews Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sellers/{seller_id}": {
      "get": {
        "tags": [
          "sellers"
        ],
        "summary": "Public seller profile with trust rating + dimension averages",
        "operationId": "seller_profile_api_v1_sellers__seller_id__get",
        "parameters": [
          {
            "name": "seller_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Seller Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SellerProfileOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tokens/balance": {
      "get": {
        "tags": [
          "tokens"
        ],
        "summary": "Your token balance",
        "operationId": "get_balance_api_v1_tokens_balance_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenBalanceOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/tokens/transactions": {
      "get": {
        "tags": [
          "tokens"
        ],
        "summary": "Your token transaction history (append-only ledger)",
        "operationId": "get_transactions_api_v1_tokens_transactions_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TokenTransactionOut"
                  },
                  "title": "Response Get Transactions Api V1 Tokens Transactions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/users/{user_id}/tokens": {
      "post": {
        "tags": [
          "tokens"
        ],
        "summary": "Admin token grant/deduct (moderator; audited via ledger)",
        "operationId": "admin_grant_api_v1_admin_users__user_id__tokens_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminTokenGrant"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenTransactionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AdminTokenGrant": {
        "properties": {
          "amount": {
            "type": "integer",
            "title": "Amount",
            "description": "Signed; positive grants, negative deducts. Non-zero."
          },
          "note": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "amount",
          "note"
        ],
        "title": "AdminTokenGrant"
      },
...
      "BadgeOut": {
        "properties": {
          "badge_id": {
            "type": "string",
            "title": "Badge Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "awarded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Awarded At"
          }
        },
        "type": "object",
        "required": [
          "badge_id",
          "name",
          "awarded_at"
        ],
        "title": "BadgeOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MemberRole": {
        "type": "string",
        "enum": [
          "user",
          "seller",
          "moderator"
        ],
        "title": "MemberRole"
      },
      "MembershipTier": {
        "type": "string",
        "enum": [
          "special",
          "founding",
          "standard"
        ],
        "title": "MembershipTier",
        "description": "Membership tiers (M1). Distinct from trust stages (reputation)."
      },
...
      "RoleUpdate": {
        "properties": {
          "role": {
            "$ref": "#/components/schemas/MemberRole"
          }
        },
        "type": "object",
        "required": [
          "role"
        ],
        "title": "RoleUpdate",
        "description": "Moderator promote/demote between user and seller (M2 slice 4).\n\n`moderator` is intentionally not grantable via the API."
      },
      "SellerProfileOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "seller_trust_score": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seller Trust Score"
          },
          "low_trust": {
            "type": "boolean",
            "title": "Low Trust",
            "default": false
          },
          "review_count": {
            "type": "integer",
            "title": "Review Count",
            "default": 0
          },
          "accuracy_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accuracy Pct"
          },
          "completeness_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completeness Pct"
          },
          "customer_service_avg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Service Avg"
          },
          "packaging_avg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Packaging Avg"
          },
          "recommend_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommend Pct"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "SellerProfileOut",
        "description": "Public seller profile with live per-dimension aggregates (M2 slice 4)."
      },
      "SellerReviewCreate": {
        "properties": {
          "accuracy": {
            "type": "boolean",
            "title": "Accuracy"
          },
          "order_completeness": {
            "type": "boolean",
            "title": "Order Completeness"
          },
          "customer_service": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Customer Service"
          },
          "packaging_quality": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Packaging Quality"
          },
          "overall_rating": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Overall Rating"
          },
          "would_recommend": {
            "type": "boolean",
            "title": "Would Recommend"
          },
          "product_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Id"
          },
          "proof_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proof Url"
          }
        },
        "type": "object",
        "required": [
          "accuracy",
          "order_completeness",
          "customer_service",
          "packaging_quality",
          "overall_rating",
          "would_recommend"
        ],
        "title": "SellerReviewCreate"
      },
      "SellerReviewOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "seller_review_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seller Review Id"
          },
          "seller_id": {
            "type": "string",
            "format": "uuid",
            "title": "Seller Id"
          },
          "reviewer_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Id"
          },
          "product_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Id"
          },
          "accuracy": {
            "type": "boolean",
            "title": "Accuracy"
          },
          "order_completeness": {
            "type": "boolean",
            "title": "Order Completeness"
          },
          "customer_service": {
            "type": "integer",
            "title": "Customer Service"
          },
          "packaging_quality": {
            "type": "integer",
            "title": "Packaging Quality"
          },
          "overall_rating": {
            "type": "integer",
            "title": "Overall Rating"
          },
          "would_recommend": {
            "type": "boolean",
            "title": "Would Recommend"
          },
          "proof_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proof Url"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "seller_id",
          "accuracy",
          "order_completeness",
          "customer_service",
          "packaging_quality",
          "overall_rating",
          "would_recommend",
          "created_at"
        ],
        "title": "SellerReviewOut"
      },
...
      "TokenBalanceOut": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "token_balance": {
            "type": "integer",
            "title": "Token Balance"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "token_balance"
        ],
        "title": "TokenBalanceOut"
      },
      "TokenKind": {
        "type": "string",
        "enum": [
          "earn_review_published",
          "earn_commission",
          "admin_grant",
          "admin_deduct",
          "adjustment"
        ],
        "title": "TokenKind",
        "description": "Token ledger entry kinds (M2 slice 7). Spending rules are M3."
      },
...
      "TokenTransactionOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "amount": {
            "type": "integer",
            "title": "Amount"
          },
          "balance_after": {
            "type": "integer",
            "title": "Balance After"
          },
          "kind": {
            "$ref": "#/components/schemas/TokenKind"
          },
          "ref_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ref Type"
          },
          "ref_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ref Id"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "amount",
          "balance_after",
          "kind",
          "created_at"
        ],
        "title": "TokenTransactionOut"
      },
      "UserOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "role": {
            "$ref": "#/components/schemas/MemberRole"
          },
          "membership_tier": {
            "$ref": "#/components/schemas/MembershipTier"
          },
          "reputation_score": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Reputation Score"
          },
          "trust_stage": {
            "type": "integer",
            "title": "Trust Stage"
          },
          "trust_level_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trust Level Name"
          },
          "is_suspended": {
            "type": "boolean",
            "title": "Is Suspended"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "role",
          "membership_tier",
          "reputation_score",
          "trust_stage",
          "is_suspended",
          "created_at"
        ],
        "title": "UserOut"
      },
      "UserTrustOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "trust_stage": {
            "type": "integer",
            "title": "Trust Stage"
          },
          "trust_level_name": {
            "type": "string",
            "title": "Trust Level Name"
          },
          "reputation_score": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Reputation Score"
          },
          "verified_review_count": {
            "type": "integer",
            "title": "Verified Review Count"
          },
          "helpfulness_ratio": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Helpfulness Ratio"
          },
          "badges": {
            "items": {
              "$ref": "#/components/schemas/BadgeOut"
            },
            "type": "array",
            "title": "Badges"
          }
        },
        "type": "object",
        "required": [
          "id",
          "trust_stage",
          "trust_level_name",
          "reputation_score",
          "verified_review_count",
          "helpfulness_ratio"
        ],
        "title": "UserTrustOut",
        "description": "Public trust profile (M2 slice 3). Stages move only via recompute."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}